@@ -1,63 +1,64 @@
 {
-  "name": "Test",
-  "version": "1.0.0",
-  "private": true,
-  "description": "Angular app running on Heroku inside an Excel Add In and connected to Salesforce.",
+  "name": "angular-app",
+  "version": "0.0.0",
   "scripts": {
-    "lint": "tslint ./src/**/*.ts -t verbose",
-    "start": "concurrently \"npm run serve-api\" \"npm run serve\"",
-    "serve": "ng serve --proxy-config proxy.config.json",
-    "serve-api": "nodemon server.js --watch server",
+    "ng": "ng",
+    "build": "ng build",
     "test": "ng test",
-    "build": "ng build --prod",
+    "lint": "ng lint",
     "e2e": "ng e2e",
-    "ng": "ng"
+    "start": "npm run build && node src/server/index.js",
+    "dev": "concurrently \"npm run server\" \"npm run client\"",
+    "server": "nodemon src/server/index.js",
+    "client": "ng serve --proxy-config proxy.config.json",
+    "postinstall": "ng build --aot -prod"
   },
-  "keywords": [],
-  "author": "Matthew Rumschlag",
-  "license": "MIT",
+  "private": true,
   "dependencies": {
-    "@angular/animations": "^7.0.0",
-    "@angular/common": "^7.0.0",
-    "@angular/compiler": "^7.0.0",
-    "@angular/core": "^7.0.0",
-    "@angular/forms": "^7.0.0",
-    "@angular/http": "^7.0.0",
-    "@angular/platform-browser": "^7.0.0",
-    "@angular/platform-browser-dynamic": "^7.0.0",
-    "@angular/router": "^7.0.0",
-    "@angular/upgrade": "^7.0.0",
-    "angular-in-memory-web-api": "^0.6.0",
+    "@angular/animations": "~7.2.0",
+    "@angular/common": "~7.2.0",
+    "@angular/compiler": "~7.2.0",
+    "@angular/core": "~7.2.0",
+    "@angular/forms": "~7.2.0",
+    "@angular/platform-browser": "~7.2.0",
+    "@angular/platform-browser-dynamic": "^7.2.0",
+    "@angular/router": "~7.2.0",
+    "@microsoft/office-js-helpers": "^1.0.2",
+    "@types/socket.io": "^2.1.2",
+    "angular-webstorage-service": "^1.0.2",
     "core-js": "^2.5.4",
+    "ejs": "^2.6.1",
     "express": "^4.16.4",
     "jsforce": "^1.9.1",
-    "ngx-socket-io": "^2.1.1",
-    "rxjs": "^6.3.0",
+    "office-toolbox": "^0.1.0",
+    "office-ui-fabric-js": "^1.5.0",
+    "rxjs": "~6.3.3",
+    "socket.io": "^2.2.0",
+    "tslib": "^1.9.0",
     "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.10.0",
-    "@angular/cli": "^7.0.0",
-    "@angular/compiler-cli": "^7.0.0",
-    "@angular/platform-server": "^7.0.0",
+    "@angular-devkit/build-angular": "~0.12.0",
+    "@angular/cli": "~7.2.2",
+    "@angular/compiler-cli": "~7.2.0",
+    "@angular/language-service": "~7.2.0",
+    "@types/body-parser": "^1.17.0",
     "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "^2.0.4",
+    "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
+    "@types/office-js": "^0.0.140",
+    "codelyzer": "~4.5.0",
     "concurrently": "^4.1.0",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.0.0",
+    "karma": "~3.1.1",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~2.0.1",
     "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "karma-phantomjs-launcher": "^1.0.2",
-    "lodash": "^4.16.2",
-    "phantomjs-prebuilt": "^2.1.7",
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.1.1"
-  },
-  "repository": {}
+    "typescript": "~3.2.2"
+  }
 }
